Match removeItem and handleItemSickness

+++ include/dw/item.h
@@ -29,6 +29,6 @@ typedef struct {
 extern Item ITEM_PARA[];
 
 int32_t giveItem(uint32_t type, uint8_t amount);
-void removeItem(uint32_t type, uint8_t amount);
+void removeItem(int32_t type, uint32_t amount);
 
 #endif
